requirements.txt
pandas
Pillow
python-docx
requests
Flask